Restructure order validation modules (#3405)
Split logic into modules based on workflow
