+++ cg/services/order_validation_service/models/order_case.py
@@ -4,11 +4,11 @@ from typing_extensions import Annotated
 from cg.constants import DataDelivery
 from cg.constants.priority import PriorityTerms
 from cg.models.orders.sample_base import NAME_PATTERN
-from cg.services.validation_service.models.case_validators import (
+from cg.services.order_validation_service.validators.case_validators import (
     validate_fathers,
     validate_mothers,
 )
-from cg.services.validation_service.models.order_sample import OrderSample
+from cg.services.order_validation_service.models.order_sample import OrderSample
 
 
 class OrderCase(BaseModel):
