+++ cg/services/order_validation_service/order_validation_service.py
@@ -1,0 +1,8 @@
+from workflows.tomte.validation_service import TomteOrderValidationService
+
+
+class OrderValidationService:
+    """Service to orchestrate the order validation."""
+
+    def __init__(self, tomte_service: TomteOrderValidationService):
+        self.tomte_service = tomte_service
