+++ cg/services/order_validation_service/validators/case_validators.py
@@ -1,8 +1,8 @@
 from pydantic import ValidationError
 from pydantic_core import InitErrorDetails, PydanticCustomError
 
-from cg.services.validation_service.models.order_case import OrderCase
-from cg.services.validation_service.models.order_sample import OrderSample
+from cg.services.order_validation_service.models.order_case import OrderCase
+from cg.services.order_validation_service.models.order_sample import OrderSample
 
 
 def validate_mothers(samples: list[OrderSample]):
