+++ cg/services/order_validation_service/workflows/tomte/models/case.py
@@ -2,8 +2,8 @@ from pydantic import AfterValidator, model_validator
 
 from cg.constants import DataDelivery, GenePanelMasterList
 from cg.models.orders.orderform_schema import OrderCase
-from cg.services.validation_service.models.case_validators import validate_subject_id
-from cg.services.validation_service.models.tomte.tomte_case_validators import (
+from cg.services.order_validation_service.validators.case_validators import validate_subject_id
+from cg.services.order_validation_service.workflows.tomte.validators.case_validators import (
     validate_tomte_delivery_type,
 )
