+++ cg/services/order_validation_service/workflows/tomte/validation_service.py
@@ -1,0 +1,5 @@
+
+class TomteOrderValidationService:
+
+    def __init__(self):
+        pass
